refactor the app foo renamed to hello
