+++ packages/create/templates/hello-nextjs/src/actions/hello.ts
@@ -2,15 +2,15 @@
 
 import { DBOS } from "@dbos-inc/dbos-sdk";
 // import {dbos_hello, dbosWorkflow, dbosBackgroundTask} from "./dbosWorkflow";
-import { fooWorkflowClass } from "../dbos/operations";
+import { helloWorkflowClass } from "../dbos/operations";
 
 
  console.log("Hello from foo.ts");
 
 // The exported function is the entry point for the workflow
 // The function is exported and not the class because Next does not support exporting classes
-export async function fooWorkflow(userName: string) {
+export async function helloWorkflow(userName: string) {
     DBOS.logger.info("Hello from DBOS!");
-    return await fooWorkflowClass.fooDBOS(userName);
+    return await helloWorkflowClass.helloDBOS(userName);
 }
