--- packages/create/templates/hello-nextjs/src/dbos/foo.ts
@@ -1,11 +1,0 @@
-import { DBOS } from "@dbos-inc/dbos-sdk";
-
-export class fooWorkflowClass {
-
-    @DBOS.transaction()
-    static async fooDBOS(userName: string) {
-        DBOS.logger.info("Hello from DBOS foooooo!");
-        
-    }
-
-}
